+++ source/commands/clustercommand.cpp
@@ -933,7 +933,7 @@ int ClusterCommand::runOptiCluster(){
             cluster.initialize(listVectorMetric, true, initialize);
             
             long long numBins = cluster.getNumBins();
-                        long long tp, tn, fp, fn;
+            double tp, tn, fp, fn;
             vector<double> results = cluster.getStats(tp, tn, fp, fn);
             m->mothurOut("0\t0\t" + toString(cutoff) + "\t" + toString(numBins) + "\t"+ toString(cutoff) + "\t" + toString(tp) + "\t" + toString(tn) + "\t" + toString(fp) + "\t" + toString(fn) + "\t");
             outStep << "0\t0\t" + toString(cutoff) + "\t" + toString(numBins) + "\t" + toString(cutoff) + "\t" << tp << '\t' << tn << '\t' << fp << '\t' << fn << '\t';
